5/3/20 Quick Update
-implemented firearm recoil
-implemented firearm hit FX & sounds
-implemented player damaging , syncd over network

+++ Assets/Code/Items/Classes/ItemWeapon.cs
@@ -2,13 +2,13 @@
 
 public abstract class ItemWeapon : Item, IWeapon
 {
-    public float Damage { get { return _damage; } set { _damage = value; } }
+    public int Damage { get { return _damage; } set { _damage = value; } }
     public float Range { get { return _range; } set { _range = value; } }
     public AudioSource Source { get; set; }
 
     [Header("Weapon Data:")]
     [SerializeField]
-    internal float _damage;
+    internal int _damage;
     [SerializeField]
     internal float _range;
