+++ Assets/Code/Items/Interfaces/IWeapon.cs
@@ -1,6 +1,6 @@
 ﻿public interface IWeapon : IItem
 {
-    float Damage { get; set; }
+    int Damage { get; set; }
     float Range { get; set; }
     UnityEngine.AudioSource Source { get; set; }
 }
